requirements.txt
Appium-Python-Client
attrs
certifi
cffi
idna
outcome
pycparser
PySocks
python-dotenv
selenium
sniffio
sortedcontainers
trio
trio-websocket
urllib3
wsproto
